fix: fixing review comments
