--- PerformanceTesting/Prerequisites/setup/src/constants.go
@@ -1,10 +1,0 @@
-package main
-
-const STRIPE_TEST_CARD_NUMBER string = "4111111111111111"
-const STRIPE_TEST_CARD_EXPIRY_MONTH string = "03"
-const STRIPE_TEST_CARD_EXPIRY_YEAR string = "23"
-const STRIPE_TEST_CARD_CVC string = "111"
-const STRIPE_TEST_KEY string = "pk_test_qH4b4EehCxTKC9SYtKHvPsPK"
-const DC_LISTING string = "listing"
-const DC_USER string = "user"
-const DC_BOOKING_REQUESTS string = "bookingRequest"
