--- PerformanceTesting/Prerequisites/setup/src/utilities.go
@@ -1,1 +1,0 @@
-package main
